Update JPA example to 5.1.0.Final
