+++ templates/play-java-intro/README.md
@@ -1,3 +1,3 @@
-# Getting started with Play (Java)
+# Getting started with Play (Java) using JPA
 
-This project demonstrate how to create a simple CRUD application with Play. A screencast to the project exist on [playframework.com](http://playframework.com)
+This project demonstrate how to create a simple CRUD application with Play using JPA. A screencast to the project exist on [playframework.com](http://playframework.com)
