--- templates/play-java-intro/app/controllers/Application.java
@@ -1,35 +1,0 @@
-package controllers;
-
-import play.*;
-import play.mvc.*;
-import play.db.jpa.*;
-import views.html.*;
-import models.Person;
-import play.data.FormFactory;
-import javax.inject.Inject;
-import java.util.List;
-
-import static play.libs.Json.*;
-
-public class Application extends Controller {
-
-    @Inject
-    FormFactory formFactory;
-
-    public Result index() {
-        return ok(index.render());
-    }
-
-    @Transactional
-    public Result addPerson() {
-        Person person = formFactory.form(Person.class).bindFromRequest().get();
-        JPA.em().persist(person);
-        return redirect(routes.Application.index());
-    }
-
-    @Transactional(readOnly = true)
-    public Result getPersons() {
-        List<Person> persons = (List<Person>) JPA.em().createQuery("select p from Person p").getResultList();
-        return ok(toJson(persons));
-    }
-}
